requirements.txt
pandas
numpy
tqdm
tenacity
nltk
requests
google-api-python-client
python-dotenv
streamlit